Implement Hash and Eq on Sid and LocalBox
This allows using LocalBox<Sid> within HashMap and HashSets.

The Hash implementation of Sid hashes the id_authority and all
subauthorities. It should be enough to obtain a hash that works well
within hashmaps.

The Eq impl is correct: Eq simply means that the PartialEq impl is
reflexive (a == a) for all a. This is obviously correct the case for
Sids: Every sid is equal to itself. (This weirdness exists to accomodate
NaNs in floats, which are not equal to themselves).